Revert a change that's not compatible with some compilers

+++ src/pylimer_tools_cpp/utils/MCUniverseGenerator.h
@@ -1945,8 +1945,7 @@ namespace utils {
       if (dist.norm() < maxDistance || maxDistance < 0.) {
         thisWeight =
           static_cast<double>(this->remainingCrossLinkerFunctionality[to]) *
-          static_cast<double>(std::expf(static_cast<float>(
-            dist.squaredNorm() * normalisationFactorInExponential)));
+          std::exp(dist.squaredNorm() * normalisationFactorInExponential);
         if (to == from) {
           thisWeight *= this->primaryLoopProbability;
         }
